Add debug output and minor refactorings

+++ c8/stm/sync.c
@@ -176,8 +176,8 @@ static inline void cond_broadcast(enum cond_type_e ctype)
 
 /************************************************************/
 
-static uint8_t number_of_segments_in_use(void) {
-    uint8_t result = 0;
+static int number_of_segments_in_use(void) {
+    int result = 0;
     int num;
     for (num = 1; num < NB_SEGMENTS; num++) {
         if (sync_ctl.in_use1[num] > 0) {
